pyproject.toml
[project]
package-mode = false
name = "RENT_STORE"
version = "0.1.0"
description = ""
authors = [
    {name = "Gui",email = "guitoledo13@hotmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "black (>=26.1.0,<27.0.0)",
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
